Route parallel_hash slots by the dedicated fold and prepare keys once
The probe and the build scatter now derive slot ids from the
JoinSlotRouting fold (UInt8, top word bits; key8/key16 keep their
bit-identical low-bit route) instead of computing the slot map's own
hash per row and discarding it. The probe prepares the block's
JoinOnKeyColumns once in ConcurrentHashJoin::joinBlock and shares it
with the lookup through RoutedJoinResult, removing the double key
materialization; the zero-copy build scatter consumes the narrow ids
directly and the copying path widens once at the IColumn::scatter
boundary. The per-family KeyGetter dispatch machinery of the old
selector (calculateHashes, hashToSelector, selectDispatchBlock) is
deleted.

Gates: parity 636 cases OK (force-pass engaged 8/8+2x0), order OK
(all checks engaged, 03448/03711 x10), gtests 19/19. The join-selector
stateless differential (new in this mission) found one candidate-only
failure, 03567_max_joined_block_size_bytes, established pre-existing
at 5b276c5fb88 (the routed probe's slot-0-only block-size estimate);
fixed in the follow-up commit per PREREG 003.

+++ src/Interpreters/HashJoin/AmacProbe.cpp
@@ -102,7 +102,7 @@ struct AmacFindPolicy
     const Map & map0;
     const Map * const * slot_maps = nullptr; /// the zero-key sentinel path only
     const SlotMapDesc * slot_descs = nullptr;
-    const UInt64 * slot_ids = nullptr; /// null at the single-slot plan
+    const UInt8 * slot_ids = nullptr; /// null at the single-slot plan
     size_t selector_base = 0; /// the first row of a continuous-range selector
     const UInt64 * selector_indexes = nullptr; /// the data of an explicit-indexes selector
     const UInt8 * skip_data = nullptr; /// null on the fast path
@@ -228,7 +228,7 @@ void amacFindPass(
     KeyGetter & key_getter,
     const Map * const * slot_maps,
     const SlotMapDesc * slot_descs,
-    const UInt64 * slot_ids,
+    const UInt8 * slot_ids,
     size_t rows,
     size_t range_first,
     const UInt64 * selector_indexes,
